added boot from OTG support
The bootscript will now use dtype and bootdev env vars if set, otherwise
will try to autodect them.

The dtype var can be one of mmc, usb, sata and if not specified a kernel
will attempt to be loaded from boot/uImage on each and the first one succeeding
will be used.

The bootdev var will be chosen based on dtype as there is typically only one
choice per dtype with the exception of usb.  If dtype is usb, bootdev needs
to refer to the usb controller which is fsl-ehci.0 for OTG and fsl-ehci.1 for
EHCI.  If the usbdevport variable is set it is assumed to be 0 or 1 to represent
which controller is being used.  The gateworks_v2014.01 bootloader will set
usbdevport automatically depending on the usb storage device selected so it
is recommended that you update to the latest bootloader if you wish to boot
from USB OTG.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>