init: add property based abstraction for DIO gpios
Ventana boards typically have 4 digital IO's going to a connector, but the
Linux gpio's associated with these varies per board. To aid in creating a
firmware image that is board-agnostic we will set an Android property
describing the gpio for each dio as 'gpio.dio<n>'.

For example, if you wish to operate on DIO0 use the value returned from
gpio.dio0 property as the gpio in /sys/class/gpio/gpio<n>

Signed-off-by: Tim Harvey <tharvey@gateworks.com>