bootscript: add support for kitkat and 3.10.x kernel
The 3.10.x kernel does not use block platform device support so we need to
specify the bootdev differently.

Additionally we are currently using an initramfs built into the kernel
at this time which may move back to an initrd decoupled from the kernel
and loaded from the bootloader in the future.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>