Disabling SELinux in kiosk mode for security file injections.
In order to place the Browser app into the list of device owners
on first boot without user confirmation, two device ownership
files must be injected into the filesystem under /data/system/.
This breaks a number of default SELinux policies with the short
workaround being to disable SELinux.

To configure the build to be SELinux compatible, one would need to:
1. Remove the selinux=0 addition to bootargs in 6x_bootscript.txt
2. Properly apply security contexts in file_contexts
3. Create a policy for the files in gateworks_kiosk.te
4. Add the .te file to the policy list in BoardConfig.mk
5. Apply proper user.group ownership in the android_file_config.h
   file mentioned below.

The other modifications and new gateworks_kiosk.te file are the
beginning of the application of the above steps but further testing
and edits to ../../system/core/include/private/android_file_config.h
are required to complete the SELinux configuration.

Signed-off-by: Robert Jones <rjones@gateworks.com>