mksdcard.sh: sfdisk version compatability fixes, removed manual offset of BOOT
With newer versions of sfdisk (v2.26+) the default behavior
changed to offsetting the first partition added by 1MiB, but
needs to be explicitly stated for older versions. Newer versions
also only accept arguments in terms of Sectors (512 bytes).
Lastly, in order to avoid manually shifting the first partition,
starting locations need to be explicitly passed as the old
version of sfdisk does not properly check for overlapping.

The new section of the script accounts for the above changes in
a way that is compatible for both without version checking.

Signed-off-by: Robert Jones <rjones@gateworks.com>