Disable strict mode checking and red border flash
While our android BSP is meant to be developer friendly, it was decided
that the red border flashing caused by the eng build's default strict
mode was more distracting than helpful. Users can always remove this line
or clear the property if they wish to turn it back on for debugging.

Signed-off-by: Robert Jones <rjones@gateworks.com>